fix(types): correct CommonJS export in @fastify/otel typings
The typescript typings were exporting using the ESM style exports, but this is a CJS module, so the exports were not resolving correctly. This switches the type export to use `export = ` syntax instead which works for ESM consumers as well as CJS consumers.

Fixes https://github.com/fastify/otel/issues/24

Signed-off-by: Bret Comnes <166301+bcomnes@users.noreply.github.com>