Use http.route attribute in line with otel spec
This commit renames the existing `http.route` to `url.path`, and then
sets `http.route` only when the value is available from routeOptions.

The [OTel semantic conventions for HTTP server spans][1] state that
`http.route`:

> MUST NOT be populated when this is not supported by the HTTP server
> framework as the route attribute should have low-cardinality and the
> URI path can NOT substitute it. SHOULD include the application root if
> there is one.

When using a parameterized route, such as /users/:userId, that means the
value of the `http.route` attribute should be "/users/:userId" and not,
for example, "/users/123". The spec instead suggests the attribute
`url.path` for that value.

[1]: https://opentelemetry.io/docs/specs/semconv/http/http-spans/#http-server-span